+++ dejafu/CHANGELOG.rst
@@ -18,6 +18,8 @@ Removed
     * ``Test.DejaFu.dejafuDiscard``
     * ``Test.DejaFu.SCT.resultsSetDiscard``
     * ``Test.DejaFu.SCT.resultsSetDiscard'``
+    * ``Test.DejaFu.SCT.runSCTDiscard``
+    * ``Test.DejaFu.SCT.runSCTDiscard'``
     * ``Test.DejaFu.SCT.sctBound``
     * ``Test.DejaFu.SCT.sctBoundDiscard``
     * ``Test.DejaFu.SCT.sctUniformRandom``